@@ -1,6 +1,6 @@
 import Rebase from 're-base';
 import firebase from 'firebase/app';
-import database from 'firebase/database';
+import 'firebase/database';
 
 var app = firebase.initializeApp({
   apiKey: 'AIzaSyCETlSYkOH6Cut5Ii31THM3j5iOiHLz89Q',
@@ -9,7 +9,7 @@ var app = firebase.initializeApp({
   projectId: 'qwales1-test-fa2c0'
 });
 
-var db = database(app);
+var db = firebase.database(app);
 var base = Rebase.createClass(db);
 
 export default base;
